Version3
This version I update the SGDT Decision Level and combine it with the Unity Planning Level. Also in the Unity Planning level I add the instinct behaviour in the individual agent and monsters can avoid collision with each and interact with explorers. But the monsters' decision part and the whole scenario design will be consider in the next version. More importantly, the agent's decision part, collision planning and deadlock solve still need to be optimised in the future version.