Version6
In this version, I update the explorer and monster 3D model and add real time agent Info showing in the simulation process. But the issue about the interface between python and unity still need to work on. Another issue is developing a python code to get the simulation data and analysis it. Further more, I might consider more complex scenarios such as adding some static obstacles stuff like that.